src/data/annotate.py
#!/usr/bin/env python3
"""
Professional annotation tool with Label Studio integration.
Supports multi-class labeling, quality control, and export to YOLO format.
"""

import json
from dataclasses import asdict, dataclass
from pathlib import Path
from typing import Dict, List, Optional

import cv2
import numpy as np
import yaml


@dataclass
class ComponentClass:
    """Component class definition with metadata."""
    id: int
    name: str
    color: tuple  # BGR for OpenCV
    description: str
    shortcuts: List[str]


@dataclass
class Annotation:
    """Single annotation with metadata."""
    image_path: str
    class_id: int
    class_name: str
    bbox: tuple  # (x, y, w, h) normalized
    confidence: float = 1.0
    annotator: str = "manual"
    timestamp: str = ""


class ProfessionalAnnotationTool:
    """
    Professional annotation tool with:
    - Multi-class support
    - Keyboard shortcuts
    - Undo/redo
    - Quality control
    - Progress tracking
    - Export to multiple formats
    """
    
    def __init__(self, config_path: str):
        self.config = self.load_config(config_path)
        self.classes = self.load_classes()
        self.annotations: List[Annotation] = []
        self.current_idx = 0
        self.history = []
        
        # Setup directories
        self.input_dir = Path(self.config['input_dir'])
        self.output_dir = Path(self.config['output_dir'])
        self.output_dir.mkdir(parents=True, exist_ok=True)
        
        # Load images
        self.images = self.load_images()
        
        # Load existing annotations
        self.load_existing_annotations()
    
    def load_config(self, config_path: str) -> dict:
        """Load configuration from YAML."""
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)
    
    def load_classes(self) -> List[ComponentClass]:
        """Load component classes from config."""
        classes = []
        for i, cls_config in enumerate(self.config['classes']):
            classes.append(ComponentClass(
                id=i,
                name=cls_config['name'],
                color=tuple(cls_config['color']),
                description=cls_config.get('description', ''),
                shortcuts=cls_config.get('shortcuts', [])
            ))
        return classes
    
    def load_images(self) -> List[Path]:
        """Load all images from input directory."""
        extensions = ['.jpg', '.jpeg', '.png']
        images = []
        for ext in extensions:
            images.extend(self.input_dir.glob(f"*{ext}"))
        return sorted(images)
    
    def load_existing_annotations(self):
        """Load previously saved annotations."""
        anno_file = self.output_dir / 'annotations.json'
        if anno_file.exists():
            with open(anno_file, 'r') as f:
                data = json.load(f)
                self.annotations = [Annotation(**a) for a in data]
            print(f"📂 Loaded {len(self.annotations)} existing annotations")
    
    def save_annotations(self):
        """Save annotations to JSON."""
        anno_file = self.output_dir / 'annotations.json'
        with open(anno_file, 'w') as f:
            json.dump([asdict(a) for a in self.annotations], f, indent=2)
        print(f"💾 Saved {len(self.annotations)} annotations")
    
    def export_to_yolo(self):
        """Export annotations to YOLO format."""
        yolo_dir = self.output_dir / 'yolo'
        (yolo_dir / 'images').mkdir(parents=True, exist_ok=True)
        (yolo_dir / 'labels').mkdir(parents=True, exist_ok=True)
        
        # Group annotations by image
        img_annotations = {}
        for anno in self.annotations:
            if anno.image_path not in img_annotations:
                img_annotations[anno.image_path] = []
            img_annotations[anno.image_path].append(anno)
        
        # Write YOLO labels
        for img_path, annos in img_annotations.items():
            img_name = Path(img_path).stem
            label_path = yolo_dir / 'labels' / f"{img_name}.txt"
            
            with open(label_path, 'w') as f:
                for anno in annos:
                    x_center, y_center, w, h = anno.bbox
                    f.write(f"{anno.class_id} {x_center} {y_center} {w} {h}\n")
        
        print(f"📤 Exported {len(img_annotations)} images to YOLO format")
    
    def display_help(self):
        """Display help overlay."""
        print("\n" + "="*60)
        print("🎯 PROFESSIONAL ANNOTATION TOOL")
        print("="*60)
        print("\nClasses:")
        for cls in self.classes:
            shortcuts = " / ".join(cls.shortcuts)
            print(f"  [{shortcuts}] → {cls.name}")
        print("\nNavigation:")
        print("  [→] Next image")
        print("  [←] Previous image")
        print("  [Space] Skip image")
        print("\nActions:")
        print("  [U] Undo last annotation")
        print("  [S] Save annotations")
        print("  [E] Export to YOLO")
        print("  [H] Show this help")
        print("  [Q] Quit")
        print("="*60 + "\n")
    
    def annotate_image(self, img_path: Path, class_obj: ComponentClass):
        """Annotate single image."""
        img = cv2.imread(str(img_path))
        h, w = img.shape[:2]
        
        # For cropped images, assume the whole image is the component
        # In production, you'd implement bbox drawing
        x_center, y_center = 0.5, 0.5
        bbox_w, bbox_h = 1.0, 1.0
        
        anno = Annotation(
            image_path=str(img_path),
            class_id=class_obj.id,
            class_name=class_obj.name,
            bbox=(x_center, y_center, bbox_w, bbox_h)
        )
        
        self.annotations.append(anno)
        self.history.append(('add', anno))
    
    def run(self):
        """Main annotation loop."""
        self.display_help()
        
        while self.current_idx < len(self.images):
            img_path = self.images[self.current_idx]
            img = cv2.imread(str(img_path))
            
            if img is None:
                self.current_idx += 1
                continue
            
            # Resize for display
            display_img = self.prepare_display(img, img_path)
            
            cv2.imshow("Annotation Tool", display_img)
            key = cv2.waitKey(0) & 0xFF
            
            # Handle keyboard input
            if not self.handle_key(key, img_path):
                break
        
        cv2.destroyAllWindows()
        self.save_annotations()
        self.export_to_yolo()
        self.print_statistics()
    
    def prepare_display(self, img: np.ndarray, img_path: Path) -> np.ndarray:
        """Prepare image for display with overlays."""
        display = img.copy()
        h, w = display.shape[:2]
        
        # Resize if too large
        max_height = 800
        if h > max_height:
            scale = max_height / h
            display = cv2.resize(display, (int(w * scale), int(h * scale)))
        
        # Add progress bar
        progress = self.current_idx / len(self.images)
        bar_height = 30
        cv2.rectangle(display, (0, 0), (int(display.shape[1] * progress), bar_height), 
                     (0, 255, 0), -1)
        
        # Add text info
        info = f"[{self.current_idx + 1}/{len(self.images)}] {img_path.name}"
        cv2.putText(display, info, (10, bar_height + 25), 
                   cv2.FONT_HERSHEY_SIMPLEX, 0.6, (255, 255, 255), 2)
        
        # Add class legend
        y_offset = bar_height + 60
        for cls in self.classes:
            shortcuts = "/".join(cls.shortcuts)
            text = f"[{shortcuts}] {cls.name}"
            cv2.putText(display, text, (10, y_offset), 
                       cv2.FONT_HERSHEY_SIMPLEX, 0.5, cls.color, 1)
            y_offset += 25
        
        return display
    
    def handle_key(self, key: int, img_path: Path) -> bool:
        """Handle keyboard input. Returns False to quit."""
        key_char = chr(key) if key < 128 else None
        
        # Check class shortcuts
        for cls in self.classes:
            if key_char in cls.shortcuts or str(cls.id) == key_char:
                self.annotate_image(img_path, cls)
                print(f"✅ [{self.current_idx + 1}] {img_path.name} → {cls.name}")
                self.current_idx += 1
                return True
        
        # Navigation and actions
        if key == 83:  # Right arrow
            self.current_idx = min(self.current_idx + 1, len(self.images) - 1)
        elif key == 81:  # Left arrow
            self.current_idx = max(self.current_idx - 1, 0)
        elif key == 32:  # Space - skip
            print(f"⏭️  [{self.current_idx + 1}] Skipped")
            self.current_idx += 1
        elif key_char == 'u':  # Undo
            self.undo()
        elif key_char == 's':  # Save
            self.save_annotations()
        elif key_char == 'e':  # Export
            self.export_to_yolo()
        elif key_char == 'h':  # Help
            self.display_help()
        elif key_char == 'q':  # Quit
            return False
        
        return True
    
    def undo(self):
        """Undo last annotation."""
        if self.history:
            action, anno = self.history.pop()
            if action == 'add' and anno in self.annotations:
                self.annotations.remove(anno)
                print(f"↶  Undone: {anno.image_path} - {anno.class_name}")
    
    def print_statistics(self):
        """Print annotation statistics."""
        print("\n" + "="*60)
        print("📊 ANNOTATION STATISTICS")
        print("="*60)
        
        # Count by class
        class_counts = {cls.name: 0 for cls in self.classes}
        for anno in self.annotations:
            class_counts[anno.class_name] += 1
        
        total = len(self.annotations)
        for cls_name, count in class_counts.items():
            percentage = (count / total * 100) if total > 0 else 0
            print(f"{cls_name:20s}: {count:4d} ({percentage:5.1f}%)")
        
        print(f"{'TOTAL':20s}: {total:4d}")
        print("="*60 + "\n")


def main():
    """Main entry point."""
    import argparse
    
    parser = argparse.ArgumentParser(description="Professional Annotation Tool")
    parser.add_argument('--config', default='configs/annotation_config.yaml',
                       help='Path to configuration file')
    args = parser.parse_args()
    
    tool = ProfessionalAnnotationTool(args.config)
    tool.run()


if __name__ == "__main__":
    main()