requirements.txt
<<<<<<< HEAD
opencv-python
numpy
=======
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
>>>>>>> b35c0bd (second commit)
